requirements.txt
click
pyserial
intelhex
pyusb
